@@ -1,7 +1,7 @@
 {
   "name": "translationCore",
   "productName": "translationCore",
-  "version": "3.2.0",
+  "version": "3.2.0-e17",
   "minCompatibleVersion": "3.2.0",
   "manifestVersion": "7",
   "description": "A bridge between TS and TM",
@@ -82,8 +82,8 @@
     "cross-env": "7.0.2",
     "css-loader": "3.4.2",
     "electron-devtools-installer": "^3.2.0",
-    "electronite": "^12.0.5-graphite",
-    "electronite-packager": "15.0.0",
+    "electronite": "file:electronite-17.3.1-graphite.tgz",
+    "electronite-packager": "file:electronite-packager-15.4.0-graphite-beta.tgz",
     "enzyme": "3.8.0",
     "enzyme-adapter-react-16": "1.8.0",
     "enzyme-to-json": "3.3.5",
@@ -117,9 +117,9 @@
   },
   "dependencies": {
     "@craco/craco": "5.6.4",
-    "@electron/remote": "1.2.0",
-    "@material-ui/core": "4.10.2",
-    "@material-ui/icons": "3.0.2",
+    "@electron/remote": "2.0.7",
+    "@material-ui/core": "4.12.3",
+    "@material-ui/icons": "4.11.2",
     "@neutrinog/electron-dl": "1.11.1",
     "@sendgrid/mail": "6.4.0",
     "adm-zip": "0.4.13",
@@ -127,7 +127,7 @@
     "archiver": "3.0.0",
     "axios": "0.19.1",
     "babel-polyfill": "6.26.0",
-    "checking-tool-wrapper": "7.0.5",
+    "checking-tool-wrapper": "7.1.0-alpha.2",
     "crypto-js": "3.1.8",
     "csv": "1.2.1",
     "deep-equal": "1.0.1",
@@ -149,7 +149,7 @@
     "ospath": "1.2.2",
     "path-extra": "3.0.0",
     "prop-types": "15.7.2",
-    "react": "16.8.6",
+    "react": "16.14.0",
     "react-bootstrap": "0.32.4",
     "react-container-dimensions": "1.3.3",
     "react-dnd": "2.5.4",
@@ -158,7 +158,7 @@
     "react-localize-redux": "3.5.3",
     "react-redux": "5.1.1",
     "react-remarkable": "1.1.3",
-    "react-scripts": "3.4.1",
+    "react-scripts": "3.4.4",
     "react-select": "1.2.0",
     "react-spotlight-clickable": "1.0.5",
     "react-toggle": "4.0.2",
@@ -178,11 +178,11 @@
     "simple-git": "1.43.0",
     "string-punctuation-tokenizer": "^2.2.0",
     "sudo-prompt": "6.2.1",
-    "tc-electron-env": "0.9.0",
+    "tc-electron-env": "0.10.0",
     "tc-source-content-updater": "1.2.10",
     "tc-strings": "0.1.7",
     "tc-tool": "4.1.0",
-    "tc-ui-toolkit": "6.0.4",
+    "tc-ui-toolkit": "6.0.5",
     "truncate-utf8-bytes": "1.0.2",
     "tsv-groupdata-parser": "0.10.3",
     "usfm-js": "3.3.1",
